+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+* Fixed an issue where clicking a mod icon in the Loadout Optimizer would select more than one of the mod.
+
 ## 6.89.0 <span class="changelog-date">(2021-10-31)</span>
 
 ## 6.88.1 <span class="changelog-date">(2021-10-28)</span>